+++ app/src/main/java/com/hapley/pocketqr/features/barcode/ui/scanner/BarcodeScannerViewModel.kt
@@ -3,17 +3,10 @@ package com.hapley.pocketqr.features.barcode.ui.scanner
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.hapley.pocketqr.features.barcode.domain.Barcode
+import com.hapley.pocketqr.features.barcode.domain.BarcodeFormat
 import com.hapley.pocketqr.features.barcode.domain.BarcodeType
 import com.hapley.pocketqr.features.barcode.domain.BarcodeUseCase
-import com.hapley.pocketqr.features.barcode.domain.CONTACT
-import com.hapley.pocketqr.features.barcode.domain.EMAIL
-import com.hapley.pocketqr.features.barcode.domain.GEO
-import com.hapley.pocketqr.features.barcode.domain.ISBN
-import com.hapley.pocketqr.features.barcode.domain.PHONE
-import com.hapley.pocketqr.features.barcode.domain.SMS
-import com.hapley.pocketqr.features.barcode.domain.UNKNOWN
-import com.hapley.pocketqr.features.barcode.domain.URL
-import com.hapley.pocketqr.features.barcode.domain.WIFI
+import com.hapley.pocketqr.features.barcode.domain.*
 import com.hapley.pocketqr.util.PocketQrUtil
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.Dispatchers
@@ -29,7 +22,7 @@ class BarcodeScannerViewModel @Inject constructor(private val barcodeUseCase: Ba
 
     fun insertBarcode(mlKitBarcode: MlKitBarcode) {
         viewModelScope.launch(Dispatchers.Main) {
-            tempRawValue = mlKitBarcode.rawValue.toString()
+            tempRawValue = mlKitBarcode.rawValue?.toString()
 
             viewModelScope.launch(Dispatchers.IO) {
                 val availableId = barcodeUseCase.getLastId() + 1
@@ -47,7 +40,7 @@ class BarcodeScannerViewModel @Inject constructor(private val barcodeUseCase: Ba
                 label = pocketQrUtil.extractSafeEntryLabel(it),
                 displayValue = mlKitBarcode.displayValue ?: "",
                 created = Date().time,
-                format = mlKitBarcode.format,
+                format = mlKitBarcode.generateFormat(),
                 type = mlKitBarcode.generateType(),
                 isFavorite = false,
                 clickCount = 0
@@ -67,4 +60,22 @@ fun MlKitBarcode.generateType(): Int = when (valueType) {
     MlKitBarcode.TYPE_URL -> URL
     MlKitBarcode.TYPE_WIFI -> WIFI
     else -> UNKNOWN
+}
+
+@BarcodeFormat
+fun MlKitBarcode.generateFormat(): Int = when (format) {
+    MlKitBarcode.FORMAT_CODE_128 -> CODE_128
+    MlKitBarcode.FORMAT_CODE_39 -> CODE_39
+    MlKitBarcode.FORMAT_CODE_93 -> CODE_93
+    MlKitBarcode.FORMAT_CODABAR -> CODEBAR
+    MlKitBarcode.FORMAT_DATA_MATRIX -> DATA_MATRIX
+    MlKitBarcode.FORMAT_EAN_13 -> EAN_13
+    MlKitBarcode.FORMAT_EAN_8 -> EAN_8
+    MlKitBarcode.FORMAT_ITF -> ITF
+    MlKitBarcode.FORMAT_QR_CODE -> QR_CODE
+    MlKitBarcode.FORMAT_UPC_A -> UPC_A
+    MlKitBarcode.FORMAT_UPC_E -> UPC_E
+    MlKitBarcode.FORMAT_PDF417 -> PDF_417
+    MlKitBarcode.FORMAT_AZTEC -> AZTEC
+    else -> FORMAT_UNKNOWN
 }