+++ preview/src/main/java/com/hapley/preview/ui/PreviewItem.kt
@@ -2,6 +2,7 @@ package com.hapley.preview.ui
 
 import android.os.Parcelable
 import kotlinx.android.parcel.Parcelize
+import com.google.zxing.BarcodeFormat as ZxingFormat
 
 @Parcelize
-data class PreviewItem (val label: String, val rawValue: String) : Parcelable
+data class PreviewItem(val label: String, val rawValue: String, val zxingFormat: ZxingFormat, val ratio: Pair<Float, Float>) : Parcelable